Move remaining env vars to next-config package
