+++ apps/app/env.ts
@@ -3,8 +3,8 @@ import { keys as auth } from '@repo/auth/keys';
 import { keys as collaboration } from '@repo/collaboration/keys';
 import { keys as database } from '@repo/database/keys';
 import { keys as email } from '@repo/email/keys';
-import { keys as core } from '@repo/env';
 import { keys as flags } from '@repo/feature-flags/keys';
+import { keys as core } from '@repo/next-config/keys';
 import { keys as observability } from '@repo/observability/keys';
 import { keys as security } from '@repo/security/keys';
 import { keys as webhooks } from '@repo/webhooks/keys';
