+++ apps/web/env.ts
@@ -1,7 +1,7 @@
 import { keys as cms } from '@repo/cms/keys';
 import { keys as email } from '@repo/email/keys';
-import { keys as core } from '@repo/env';
 import { keys as flags } from '@repo/feature-flags/keys';
+import { keys as core } from '@repo/next-config/keys';
 import { keys as observability } from '@repo/observability/keys';
 import { keys as rateLimit } from '@repo/rate-limit/keys';
 import { keys as security } from '@repo/security/keys';
