+++ apps/web/package.json
@@ -15,7 +15,6 @@
     "@repo/cms": "workspace:*",
     "@repo/design-system": "workspace:*",
     "@repo/email": "workspace:*",
-    "@repo/env": "workspace:*",
     "@repo/feature-flags": "workspace:*",
     "@repo/next-config": "workspace:*",
     "@repo/observability": "workspace:*",
