+++ packages/next-config/package.json
@@ -7,12 +7,14 @@
     "typecheck": "tsc --noEmit --emitDeclarationOnly false"
   },
   "devDependencies": {
-    "@repo/env": "workspace:*",
     "@repo/typescript-config": "workspace:*",
     "next": "15.1.0"
   },
   "dependencies": {
     "@next/bundle-analyzer": "^15.1.0",
-    "@prisma/nextjs-monorepo-workaround-plugin": "^6.0.1"
+    "@prisma/nextjs-monorepo-workaround-plugin": "^6.0.1",
+    "@t3-oss/env-core": "^0.11.1",
+    "@t3-oss/env-nextjs": "^0.11.1",
+    "zod": "^3.24.1"
   }
 }
